Benchmark formatting minor fixes.

+++ test/perft.cpp
@@ -65,7 +65,7 @@ int main(int argv, char** argc){
     perft(depth);
     std::chrono::steady_clock::time_point end_time(std::chrono::steady_clock::now());
 
-    uint ms = std::chrono::duration_cast<std::chrono::milliseconds>(end_time-start_time).count();
+    ulong ms = std::chrono::duration_cast<std::chrono::milliseconds>(end_time-start_time).count();
     std::cout << "time: " << ms << " ms" << std::endl;
     std::cout << "perft: " << leaves_count << std::endl;
     std::cout << "number of states: " << states_count << " (" << std::fixed << count_per_sec(states_count, ms) << " states/sec)" << std::endl;
